some smal bug fixes, now loading only accounts on the main activity, and saving imapPort correctly

+++ app/src/main/java/tech/overturn/service/Fetcher.java
@@ -68,7 +68,7 @@ public class Fetcher {
         return new URLName(
                 protocol,
                 a.imapHost,
-                a.imapPort,
+                a.imapPort.intValue(),
                 "Inbox",
                 a.user,
                 a.password);
